+++ lib/worker_manager.dart
@@ -3,4 +3,8 @@
 // found in the LICENSE file.
 library worker_manager;
 
+export 'src/cancelable.dart';
 export 'src/executor.dart';
+export 'src/isolate_wrapper.dart';
+export 'src/runnable.dart';
+export 'src/task.dart';